+++ argbind/argbind.py
@@ -60,15 +60,30 @@ def _format_func_debug(func_name, func_kwargs, scope=None):
     formatted.append(")")
     return '\n'.join(formatted)
 
-def bind(*patterns, without_prefix=False):
-    """
-    Wrap the function so it looks in ARGS (managed 
-    by the scope context manager) for keyword 
-    arguments.
+def bind(*patterns, without_prefix=False, positional=False):
+    """Binds a functions arguments so that it looks up argument
+    values in a dictionary scoped by ArgBind.
+
+    Parameters
+    ----------
+    without_prefix : bool, optional
+        Whether or not to bind without the function name as the prefix. 
+        If True, the functions arguments will be available at "arg_name"
+        rather than "func_name.arg_name", by default False
+    positional : bool, optional
+        Arguments that are not keyword arguments are not bound by default. If
+        this is True, then the arguments will be bound as positional arguments
+        in some order, by default False
     """
 
+    if positional and patterns:
+        warnings.warn(
+            f"Combining positional arguments with scoping patterns is not allowed. Removing scoping patterns {patterns}. \n"
+            "See https://github.com/pseeth/argbind/tree/main/examples/hello_world#argbind-with-positional-arguments")
+        patterns = []
+
     def decorator(func):
-        PARSE_FUNCS[func.__name__] = (func, patterns, without_prefix)
+        PARSE_FUNCS[func.__name__] = (func, patterns, without_prefix, positional)
         
         @wraps(func)
         def cmd_func(*args, **kwargs):
@@ -79,7 +94,7 @@ def bind(*patterns, without_prefix=False):
             for key, val in sig.parameters.items():
                 arg_type = val.annotation
                 arg_val = val.default
-                if arg_val is not inspect.Parameter.empty:
+                if arg_val is not inspect.Parameter.empty or positional:
                     arg_name = f'{prefix}.{key}' if not without_prefix else f'{key}'
                     if arg_name in ARGS and key not in kwargs:
                         cmd_kwargs[key] = ARGS[arg_name]
@@ -87,7 +102,7 @@ def bind(*patterns, without_prefix=False):
                         if PATTERN:
                             use_key = f'{PATTERN}/{use_key}'
                         USED_ARGS[use_key] = ARGS[arg_name]
-            
+
             kwargs.update(cmd_kwargs)
             if 'args.debug' not in ARGS: ARGS['args.debug'] = False
             if ARGS['args.debug'] or DEBUG:
@@ -241,7 +256,7 @@ def parse_args():
 
     # Add kwargs from function to parser
     for func_name in PARSE_FUNCS:
-        func, patterns, without_prefix = PARSE_FUNCS[func_name]
+        func, patterns, without_prefix, positional = PARSE_FUNCS[func_name]
         sig = inspect.signature(func)
         prefix = func.__name__
 
@@ -255,27 +270,42 @@ def parse_args():
             title=f"Generated arguments for function {prefix}",
         )
 
+        def _get_arg_names(key, is_kwarg):
+            arg_names = []
+            arg_name = key
+
+            prepend = '--' if is_kwarg else ''
+            if without_prefix:
+                arg_name = prepend + f'PATTERN/{key}'
+            else:
+                arg_name = prepend + f'PATTERN/{prefix}.{key}'
+
+            arg_names.append(arg_name.replace('PATTERN/', ''))
+            
+            if patterns is not None:
+                for p in patterns:
+                    arg_names.append(
+                        arg_name.replace('PATTERN', p)
+                    )
+            return arg_names
+
+
         for key, val in sig.parameters.items():
             arg_type = val.annotation
             arg_val = val.default
+            is_kwarg = arg_val is not inspect.Parameter.empty
 
-            if arg_val is not inspect.Parameter.empty:
-                arg_names = []
+            if is_kwarg or positional:
+                arg_names = _get_arg_names(key, is_kwarg)
                 arg_help = {}
                 help_text = ''
                 if key in parameter_help:
                     help_text = textwrap.fill(parameter_help[key], width=HELP_WIDTH)
-                if not without_prefix:
-                    arg_names.append(f'--{prefix}.{key}')
-                else:
-                    arg_names.append(f'--{key}')
-                arg_help[arg_names[-1]] = help_text
-                for pattern in patterns:
-                    if not without_prefix:
-                        arg_names.append(f'--{pattern}/{prefix}.{key}')
-                    else:
-                        arg_names.append(f'--{pattern}/{key}')
-                    arg_help[arg_names[-1]] = argparse.SUPPRESS
+                arg_help[arg_names[0]] = help_text
+                if len(arg_names) > 1:
+                    for pattern_arg_name in arg_names[1:]:
+                        arg_help[pattern_arg_name] = argparse.SUPPRESS
+
                 for arg_name in arg_names:
                     inner_types = [str, int, float, bool]
                     list_types = [List[x] for x in inner_types]