+++ changelog.md
@@ -1,4 +1,8 @@
 # Changelog
+## v0.1.9
+- Positional arguments can now be bound with `positional=True`. ArgBind should now be able to build programs
+  with identical APIs to ArgParse, with less code and added support for .yaml files!
+
 ## v0.1.8
 - Environment variables can now be referenced within YAML files. All variables that are in `os.environ` are used to resolve any values that start with `$` in a YAML file.
 - Variables now resolve not only for strings but also within lists of strings.